Fixed test cases related to time and timestamp to use the timezone.

+++ java/adapter/jdbc/src/test/java/org/apache/arrow/adapter/jdbc/JdbcToArrowTestHelper.java
@@ -151,7 +151,7 @@ public class JdbcToArrowTestHelper {
         return true;
     }
 
-    public static boolean assertTimeVectorValues(FieldVector fx, int rowCount, int[] values){
+    public static boolean assertTimeVectorValues(FieldVector fx, int rowCount, long[] values){
         TimeMilliVector timeMilliVector = ((TimeMilliVector)fx);
 
         assertEquals(rowCount, timeMilliVector.getValueCount());
